feat: refactoring load_kubernetes_data to kubernetes_manager module
