+++ app/v1/modules/kubernetes_manager/__init__.py
@@ -8,6 +8,7 @@ from .script_create_vars_file import ScriptVarsFileCreator
 from .script_create_workload_template import ScriptWorkloadTemplateCreator
 from .script_create_workload import ScriptWorkloadCreator
 from .script_load_kubernetes_config import ScriptKubernetesConfigLoader
+from .script_load_kubernetes_data import ScriptKubernetesDataLoader
 
 
 __all__ = [
@@ -18,4 +19,5 @@ __all__ = [
     "ScriptWorkloadTemplateCreator",
     "ScriptWorkloadCreator",
     "ScriptKubernetesConfigLoader",
+    "ScriptKubernetesDataLoader",
 ]
